requirements.txt
boto3==1.14.60
Flask==1.1.2
Flask-Cors==3.0.9
flask-praetorian==1.2.0
Flask-SQLAlchemy==2.4.4
google-auth==1.30.1
mysql-connector-python==8.0.21
PyJWT==2.1.0
pybase64==1.1.4
python-decouple==3.4
requests==2.25.1
SQLAlchemy==1.3.20
subprocess.run==0.0.8
uuid==1.30
nltk==3.6.2